requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
cffi==1.14.5
click==8.0.0a1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2b1
django-braces==1.14.0
filelock==3.0.12
flower==0.9.7
gunicorn==20.0.4
humanize==3.2.0
kombu==4.6.11
objects==0.3.1
Pillow==8.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2021.1
rabbitmq==0.2.0
redis==3.5.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
sync==1.0.0
tornado==6.1
vine==1.3.0
virtualenv==20.4.2
wcwidth==0.2.5
whitenoise==5.2.0
